Stats: Ensure test is run as integration test
